+++ js/index.js
@@ -2,11 +2,10 @@
 
 import selectHandler from "./select-handler.js";
 import {createMovieList, populateMovieInfo} from "./edit-movies.js";
-// import getMovieCover from "./get-movie-cover.js";
-import MV_KEY from "./keys.js";
+import {renderMovie, renderMovieInitial} from "./render-movies.js";
 
 
-export {allMovies, renderMovie, movieForm};
+export {allMovies, renderMovie, movieForm, movieContainer, loading};
 
 // constants************************************************************************
 const movieContainer = document.querySelector("#movie-container");
@@ -14,78 +13,25 @@ const submitBtn = document.querySelector("#submit-btn");
 const title = document.querySelector("#title");
 const rating = document.querySelector("#rating");
 const movieForm = document.querySelector("#movie-form");
-const movieSelect = document.querySelector("#movie-select")
+const movieSelect = document.querySelector("#movie-select");
+const loading = document.querySelector(".loading");
 
 // functions************************************************************************
 // fetches movies.json while 'loading'. Sends response to be rendered---------------
 async function allMovies(movieList) {
+        loading.classList.remove("d-none");
     if (movieList === undefined) {
-        movieContainer.innerHTML = `<h2>Loading...</h2>`;
+        console.log("undefined movielist");
         const response = await fetch("http://localhost:3000/movies").then(resp => resp.json()).catch(error => console.log("Error", error));
-        movieContainer.innerHTML = ``;
-        renderMovie(response);
+        loading.classList.add("d-none");
+        renderMovieInitial(response);
     } else {
-        movieContainer.innerHTML = `<h2>Loading...</h2>`;
-        movieContainer.innerHTML = ``;
+        const loading = document.createElement("div");
+        loading.classList.add("loading");
+        movieContainer.appendChild(loading);
         renderMovie(movieList);
+        loading.classList.add("d-none");
     }
-}
-
-// renders movies onto DOM----------------------------------------------------------
-function renderMovie(movies) {
-    movies.forEach(movie => {
-        const movieCardSpacer = document.createElement("div")
-        const movieCard = document.createElement("div")
-        const movieTitle = document.createElement("h3")
-        const movieRating = document.createElement("div")
-        const deleteBtn = document.createElement("button")
-        movieTitle.innerHTML = movie.title;
-        movieRating.innerHTML = addStars(movie.rating);
-        deleteBtn.innerText = "X";
-        movieCard.classList.add("movie-card");
-        movieCardSpacer.classList.add("movie-card-spacer");
-        deleteBtn.classList.add("delete-btn");
-        deleteBtn.style.display = "none";
-        movieCard.appendChild(movieTitle);
-        movieCard.appendChild(movieRating);
-        movieCard.appendChild(deleteBtn);
-        movieCard.addEventListener("mouseenter", showDelete);
-        movieCard.addEventListener("mouseleave", hideDelete);
-        movieCardSpacer.appendChild(movieCard);
-        movieContainer.appendChild(movieCardSpacer);
-        deleteBtn.addEventListener("click", async function () {
-            try {
-                const url = `http://localhost:3000/movies/${movie.id}`; // to keep movie.id, we chose not to use callback fn.
-                const options = {
-                    method: "DELETE",
-                    headers: {
-                        'Content-Type': 'application/json'
-                    }
-                }
-                const resp = await fetch(url, options)
-                    .catch(error => console.log("error" + error));
-                await allMovies();
-            } catch (error) {
-                console.error(error);
-            }
-        })
-    })
-}
-
-function addStars(rating) {
-    let stars = "";
-    for (let i = 0; i < rating; i++) {
-        stars += `<i class="fa-solid fa-star" style="color: #000000;"></i>`;
-    }
-    return stars;
-}
-
-function showDelete() {
-    this.lastElementChild.style.display = "block"
-}
-
-function hideDelete() {
-    this.lastElementChild.style.display = "none"
 }
 
 //Determines source of function call, sends to select-handler module----------------
