+++ js/sort-movies.js
@@ -1,29 +1,41 @@
 "use strict";
 
-import {allMovies} from "./index.js";
+import {allMovies, loading} from "./index.js";
 
 function updateRating(movieList) {
-    const filteredMovies = movieList.filter((movie) => {
-        return movie.rating.toString() === document.querySelector("#rating").value;
-    });
-    allMovies(filteredMovies);
+    // document.querySelector("#title").value = "";
+    if (document.querySelector("#rating").value !== "all") {
+        const filteredMovies = movieList.filter((movie) => {
+            return movie.rating.toString() === document.querySelector("#rating").value;
+        });
+        allMovies(filteredMovies);
+    } else {
+        allMovies();
+    }
 }
 
-
-function sortByTitle(movieList){
-         const filteredMovies = movieList.filter((movie) => {
-         const movieCards = document.querySelectorAll(".movie-card")
-             movieCards.forEach(movie => movie.classList.remove("movie-card-anime"))
-             return movie.title.toLowerCase().includes(document.querySelector("#title").value.toLowerCase().trim())
-    });
-    allMovies(filteredMovies);
+function updateTitle(movieList) {
+    if (document.querySelector("#rating").value !== "all") {
+        let filteredMoviesByRating = movieList.filter((movie) => {
+            return movie.rating.toString() === document.querySelector("#rating").value;
+        });
+        const filteredMovies = filteredMoviesByRating.filter((movie) => {
+            return movie.title.toLowerCase().includes(document.querySelector("#title").value.toLowerCase().trim());
+        });
+            allMovies(filteredMovies);
+    } else {
+        const filteredMovies = movieList.filter((movie) => {
+            return movie.title.toLowerCase().includes(document.querySelector("#title").value.toLowerCase().trim());
+        });
+        allMovies(filteredMovies);
+    }
 }
 
-// sets functionality of form inputs to filtering by title/rating---------------
+// sets functionality of form inputs to filtering by title/rating-------------------
 export default function sortMovies(e, movieList) {
     if (e.target.id === "rating") {
         updateRating(movieList);
     } else {
-        sortByTitle(movieList);
+        updateTitle(movieList);
     }
 }
